# src/main/java/com/reservation/app/Controller/EventController.java
package com.reservation.app.Controller;
import org.springframework.web.bind.annotation.*;


@RestController
@RequestMapping("/events")
public class EventController {

    @GetMapping
    public ResponseEntity<List<Event>> getAllEvents() {
        // Code to fetch all events
    }

    @PostMapping
    @PreAuthorize("hasRole('ADMIN')")
    public ResponseEntity<Event> createEvent(@RequestBody Event event) {
        // Code to create a new event
    }

    @GetMapping("/{eventName}")
    public ResponseEntity<EventDetails> getEventDetails(@PathVariable String eventName) {
        // Code to fetch event details along with tables
    }

    @DeleteMapping("/{eventName}")
    @PreAuthorize("hasRole('ADMIN')")
    public ResponseEntity<Void> deleteEvent(@PathVariable String eventName) {
        // Code to delete event
    }
}